requirements.txt
google-generativeai>=0.3.0
streamlit>=1.24.0
python-dotenv>=1.0.0
google-cloud-dialogflow-cx>=1.16.0